+++ eyeballvul/score.py
@@ -1,5 +1,8 @@
 import json
+from collections.abc import Callable
+from dataclasses import dataclass
 from datetime import datetime
+from typing import Any, Literal, cast
 
 from litellm import completion, supports_function_calling
 from sqlmodel import Session, create_engine, select
@@ -7,6 +10,7 @@ from typeguard import typechecked
 
 from eyeballvul.config.config_loader import Config
 from eyeballvul.models.eyeballvul import EyeballvulItem
+from eyeballvul.util import get_str_weak_hash
 
 instructions_template = """
 An AI vulnerability detection tool has analyzed a project and come up with the following lead:
@@ -28,127 +32,102 @@ The real vulnerabilities are as follows:
 
 Use the tool "return_score" to return your response.
 """
+instructions_template_hash = get_str_weak_hash(instructions_template)[:20]
 
 
+@dataclass
+class Stats:
+    """
+    False negatives, true positives, and false positives.
+
+    False positives may be None if this object is part of a larger StatsWithCutoff object, as the
+    false positives apply to the whole set of vulnerabilities, not just the ones before or after a
+    certain date.
+    """
+
+    fn: int
+    tp: int
+    fp: int | None = None
+
+    def to_dict(self) -> dict[str, int | None]:
+        return {"fn": self.fn, "tp": self.tp, "fp": self.fp}
+
+
+@dataclass
 @typechecked
-def score(
-    commit_hash: str,
-    vulns_submission: list[str],
-    cutoff_date: datetime | None = None,
-    scoring_model: str = Config.scoring_model,
-) -> tuple[dict[str, int] | dict[str, int | dict[str, int]], dict[int, str]]:
+class StatsWithCutoff:
+    """Similar to `Stats`, but with separate false negative and true positive counts before and
+    after a cutoff."""
+
+    fp: int
+    before: Stats
+    after: Stats
+
+    def to_dict(self) -> dict[str, Any]:
+        return {
+            "fp": self.fp,
+            "before": self.before.to_dict(),
+            "after": self.after.to_dict(),
+        }
+
+
+class EyeballvulScore:
     """
-    Score a model's response (`vulns_submission`) against the real list of vulnerabilities at
-    `commit_hash`.
+    Representation of the result of scoring a model's attempt at finding vulnerabilities.
 
-    Returns a tuple of two dictionaries. The first one contains the score as a dictionary of true positives, false positives, etc. The second one is a mapping from vulnerability indices in the input list to the corresponding vulnerability IDs.
-
-    If `cutoff_date` is not provided, the first dictionary returned has the following format:
-    ```
-    {
-        "true_positive": int,
-        "false_positive": int,
-        "false_negative": int,
-    }
-    ```
-
-    If `cutoff_date` is provided, the first dictionary has the following format:
-    ```
-    {
-        "false_positive": int,
-        "before_cutoff": {
-            "true_positive": int,
-            "false_negative": int,
-        },
-        "after_cutoff": {
-            "true_positive": int,
-            "false_negative": int,
-        },
-    }
-    ```
-    But one of the keys ("before_cutoff", "after_cutoff") may be missing if there are no vulnerabilities before or after the cutoff date.
-
-    The second dictionary returned has the following format:
-    ```
-    {
-        <index in the input list (starts at 0)>: "<corresponding vulnerability ID>",
-    }
-
-
-    The value in `Config.scoring_model` is used by default for the scoring model, but this can be changed. The model name should be a valid LiteLLM model name, and support function calling / tool use.
+    The `stats` attribute contains the false negatives, true positives, and false positives. The `mapping` attribute is a dictionary mapping the index of each vulnerability submission to the ID of the real vulnerability it corresponds to (if any). The `vuln_dates` attribute is a dictionary mapping vulnerability IDs to their publication date. The `scoring_model` attribute is the name of the scoring model used.
     """
-    if not supports_function_calling(scoring_model):
-        raise ValueError(f"The model {scoring_model} doesn't support function calling.")
-    engine = create_engine(f"sqlite:///{Config.paths.db}/eyeballvul.db")
-    with Session(engine) as session:
-        real_vulns = list(
-            session.exec(
-                select(EyeballvulItem).where(EyeballvulItem.commits.contains(commit_hash))  # type: ignore[attr-defined]
-            ).all()
-        )
-    if cutoff_date:
-        return _score_with_cutoff(vulns_submission, real_vulns, cutoff_date, scoring_model)
-    else:
-        return _score_without_cutoff(vulns_submission, real_vulns, scoring_model)
 
+    def __init__(
+        self,
+        stats: Stats,
+        mapping: dict[int, str],
+        vuln_dates: dict[str, datetime],
+        scoring_model: str,
+        type: Literal["llm"],
+    ):
+        if not stats.fp:
+            raise ValueError(
+                "fp can't be None in the EyeballvulStats object given to EyeballvulScore."
+            )
+        self._stats = stats
+        self.mapping = mapping
+        self.vuln_dates = vuln_dates
+        self.type = type
+        self.id = f"{scoring_model}@{instructions_template_hash}"
 
-def _score_without_cutoff(
-    vulns_submission: list[str], real_vulns: list[EyeballvulItem], scoring_model: str
-) -> tuple[dict[str, int], dict[int, str]]:
-    real_vulns_hit = set()
-    total_score = 0
-    real_vuln_mapping = {}
-    for i, vuln_submission in enumerate(vulns_submission):
-        score, corresponds_to = score_one(vuln_submission, real_vulns, scoring_model)
-        if corresponds_to:
-            real_vulns_hit.add(corresponds_to)
-            real_vuln_mapping[i] = corresponds_to
-            total_score += score
-    true_positive = len(real_vulns_hit)
-    false_positive = len(vulns_submission) - total_score
-    false_negative = len(real_vulns) - true_positive
-    score_result = {
-        "true_positive": true_positive,
-        "false_positive": false_positive,
-        "false_negative": false_negative,
-    }
-    return score_result, real_vuln_mapping
+    def stats(
+        self,
+        cutoff_date: datetime | None = None,
+    ) -> Stats | StatsWithCutoff:
+        if not cutoff_date:
+            return self._stats
+        else:
+            vulns_before = {
+                vuln_id for vuln_id, date in self.vuln_dates.items() if date < cutoff_date
+            }
+            vulns_after = {
+                vuln_id for vuln_id, date in self.vuln_dates.items() if date >= cutoff_date
+            }
+            vulns_hit = set(self.mapping.values())
+            tp_before = len(vulns_before & vulns_hit)
+            fn_before = len(vulns_before - vulns_hit)
+            tp_after = len(vulns_after & vulns_hit)
+            fn_after = len(vulns_after - vulns_hit)
+            return StatsWithCutoff(
+                fp=cast(int, self._stats.fp),
+                before=Stats(fn=fn_before, tp=tp_before),
+                after=Stats(fn=fn_after, tp=tp_after),
+            )
 
-
-def _score_with_cutoff(
-    vulns_submission: list[str],
-    real_vulns: list[EyeballvulItem],
-    cutoff_date: datetime,
-    scoring_model: str,
-) -> tuple[dict[str, int | dict[str, int]], dict[int, str]]:
-    real_vuln_ids = {
-        "before_cutoff": {vuln.id for vuln in real_vulns if vuln.published < cutoff_date},
-        "after_cutoff": {vuln.id for vuln in real_vulns if vuln.published >= cutoff_date},
-    }
-    real_vulns_hit: dict[str, set[str]] = {"before_cutoff": set(), "after_cutoff": set()}
-    total_score = {"before_cutoff": 0, "after_cutoff": 0}
-    real_vuln_mapping = {}
-    for i, vuln_submission in enumerate(vulns_submission):
-        score, corresponds_to = score_one(vuln_submission, real_vulns, scoring_model)
-        if corresponds_to:
-            if corresponds_to in real_vuln_ids["before_cutoff"]:
-                when = "before_cutoff"
-            else:
-                when = "after_cutoff"
-            real_vulns_hit[when].add(corresponds_to)
-            real_vuln_mapping[i] = corresponds_to
-            total_score[when] += score
-    score_result: dict[str, int | dict[str, int]] = {
-        "false_positive": len(vulns_submission)
-        - total_score["before_cutoff"]
-        - total_score["after_cutoff"],
-    }
-    for when in ("before_cutoff", "after_cutoff"):
-        if real_vuln_ids[when]:
-            true_positive = len(real_vulns_hit[when])
-            false_negative = len(real_vuln_ids[when]) - true_positive
-            score_result[when] = {"true_positive": true_positive, "false_negative": false_negative}
-    return score_result, real_vuln_mapping
+    def to_dict(self) -> dict[str, Any]:
+        return {
+            "stats": self._stats.to_dict(),
+            "mapping": self.mapping,
+            "vuln_dates": {k: v.isoformat() for k, v in self.vuln_dates.items()},
+            "type": self.type,
+            "id": self.id,
+        }
 
 
 @typechecked
@@ -206,3 +185,50 @@ def score_one(
             f"Score of 1 given but the `corresponds_to` parameter ('{arguments_dict['corresponds_to']}') is not a valid vulnerability ID. Valid IDs: {real_vuln_ids}."
         )
     return (score, arguments_dict.get("corresponds_to"))
+
+
+@typechecked
+def compute_score(
+    commit_hash: str,
+    vulns_submission: list[str],
+    scoring_model: str = Config.scoring_model,
+    score_one_fn: Callable = score_one,
+) -> EyeballvulScore:
+    """
+    Score a model's response (`vulns_submission`) against the real list of vulnerabilities at
+    `commit_hash`.
+
+    Returns an `EyeballvulScore` object. See its documentation for further information.
+
+    The value in `Config.scoring_model` is used by default for the scoring model, but this can be changed with the `scoring_model` argument. The model name should be a valid LiteLLM model name, and support function calling / tool use.
+
+    It's possible to supply a custom `score_one_fn` function, which must have the same signature as `score_one` (the one used by default).
+    """
+    if not supports_function_calling(scoring_model):
+        raise ValueError(f"The model {scoring_model} doesn't support function calling.")
+    engine = create_engine(f"sqlite:///{Config.paths.db}/eyeballvul.db")
+    with Session(engine) as session:
+        real_vulns = list(
+            session.exec(
+                select(EyeballvulItem).where(EyeballvulItem.commits.contains(commit_hash))  # type: ignore[attr-defined]
+            ).all()
+        )
+    real_vulns_hit = set()
+    total_score = 0
+    real_vuln_mapping = {}
+    for i, vuln_submission in enumerate(vulns_submission):
+        score, corresponds_to = score_one_fn(vuln_submission, real_vulns, scoring_model)
+        if corresponds_to:
+            real_vulns_hit.add(corresponds_to)
+            real_vuln_mapping[i] = corresponds_to
+            total_score += score
+    tp = len(real_vulns_hit)
+    fp = len(vulns_submission) - total_score
+    fn = len(real_vulns) - tp
+    return EyeballvulScore(
+        stats=Stats(fp=fp, fn=fn, tp=tp),
+        mapping=real_vuln_mapping,
+        vuln_dates={vuln.id: vuln.published for vuln in real_vulns},
+        type="llm",
+        scoring_model=scoring_model,
+    )